feat: improve code
